refactor: rename Password components
