+++ packages/shared/components/inputs/index.js
@@ -16,7 +16,7 @@ export { default as Input } from './Input.svelte'
 export { default as InputContainer } from './InputContainer.svelte'
 export { default as Number } from './Number.svelte'
 export { default as NumberInput } from './NumberInput.svelte'
-export { default as Password } from './Password.svelte'
+export { default as PasswordInput } from './PasswordInput.svelte'
 export { default as Pin } from './Pin.svelte'
 export { default as Radio } from './Radio.svelte'
 export { default as RecipientInput } from './RecipientInput.svelte'
